Task: restructured the hero section file to be in the home folder
