+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import HeroSection from "./ui/hero-section";
+import HeroSection from "./ui/home/hero-section";
 
 export default function Page() {
   return <HeroSection />;
