+++ src/app/ui/about-us/our-values.tsx
@@ -43,7 +43,7 @@ export default function OurValues() {
           fill="none"
           xmlns="http://www.w3.org/2000/svg"
         >
-          <g clip-path="url(#clip0_146_11592)">
+          <g clipPath="url(#clip0_146_11592)">
             <path
               d="M10 20.0111C15.5228 20.0111 20 15.5339 20 10.0111C20 4.48821 15.5228 0.0111084 10 0.0111084C4.47715 0.0111084 0 4.48821 0 10.0111C0 15.5339 4.47715 20.0111 10 20.0111Z"
               fill="#666666"
@@ -65,7 +65,7 @@ export default function OurValues() {
               fill="#141414"
             />
           </g>
-          <g clip-path="url(#clip1_146_11592)">
+          <g clipPath="url(#clip1_146_11592)">
             <path
               d="M30 16.0066C33.3137 16.0066 36 13.3203 36 10.0066C36 6.69291 33.3137 4.00665 30 4.00665C26.6863 4.00665 24 6.69291 24 10.0066C24 13.3203 26.6863 16.0066 30 16.0066Z"
               fill="#333333"
@@ -87,7 +87,7 @@ export default function OurValues() {
               fill="#141414"
             />
           </g>
-          <g clip-path="url(#clip2_146_11592)">
+          <g clipPath="url(#clip2_146_11592)">
             <path
               d="M42.8 12.8031C44.3464 12.8031 45.6 11.5495 45.6 10.0031C45.6 8.45671 44.3464 7.20312 42.8 7.20312C41.2536 7.20312 40 8.45671 40 10.0031C40 11.5495 41.2536 12.8031 42.8 12.8031Z"
               fill="#333333"
